@@ -1,9 +1,8 @@
 ﻿using System;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
+
+public abstract class ScorerBase<T> : IRatioScorer<T> where T : IEquatable<T>
 {
-    public abstract class ScorerBase<T> : IRatioScorer<T> where T : IEquatable<T>
-    {
-        public abstract int Score(T[] input1, T[] input2);
-    }
+    public abstract int Score(T[] input1, T[] input2);
 }
