@@ -1,10 +1,8 @@
-﻿using System;
-using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
+﻿using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public sealed class PartialTokenAbbreviationScorer : TokenAbbreviationScorerBase
 {
-    public class PartialTokenAbbreviationScorer : TokenAbbreviationScorerBase
-    {
-        protected override Func<string, string, int> Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
