@@ -1,9 +1,6 @@
-﻿using System;
+﻿namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+public abstract class StrategySensitiveScorerBase : ScorerBase
 {
-    public abstract class StrategySensitiveScorerBase : ScorerBase
-    {
-        protected abstract Func<string, string, int> Scorer { get; }
-    }
+    protected abstract FuzzySharp.Scorer Scorer { get; }
 }
